+++ cmd/root.go
@@ -22,29 +22,32 @@ import (
 
 var k = koanf.New(".")
 
-func Execute(version, commit, date string) {
+func Execute(version, commit, date string) int {
 	logger, _ := zap.NewProduction()
 	defer logger.Sync() //nolint:errcheck
 
 	f := config.FlagSet()
 	if err := f.Parse(os.Args[1:]); err != nil {
-		logger.Fatal(fmt.Sprintf("error parsing cli args: %v", err))
+		logger.Error(fmt.Sprintf("error parsing cli args: %v", err))
+		return 1
 	}
 
 	if versionFlag, _ := f.GetBool("version"); versionFlag {
 		fmt.Printf("version: %s\ncommit: %s\ndate: %s\ngo: %s\n", version, commit, date, runtime.Version())
-		os.Exit(0)
+		return 0
 	}
 
-	configFile, _ := f.GetString("configfile")
+	configFile, _ := f.GetString("config")
 	if configFile != "" {
 		if err := k.Load(file.Provider(configFile), yaml.Parser()); err != nil {
-			logger.Fatal(fmt.Sprintf("error loading config: %v", err))
+			logger.Error(fmt.Sprintf("error loading config: %v", err))
+			return 1
 		}
 	}
 
 	if err := k.Load(posflag.Provider(f, ".", k), nil); err != nil {
-		logger.Fatal(fmt.Sprintf("error loading config: %v", err))
+		logger.Error(fmt.Sprintf("error loading config: %v", err))
+		return 1
 	}
 
 	if err := k.Load(env.ProviderWithValue("CONFIG_", ".", func(s string, v string) (string, interface{}) {
@@ -58,17 +61,20 @@ func Execute(version, commit, date string) {
 		// Otherwise, return the plain string.
 		return key, v
 	}), nil); err != nil {
-		logger.Fatal(fmt.Sprintf("error loading config: %v", err))
+		logger.Error(fmt.Sprintf("error loading config: %v", err))
+		return 1
 	}
 
 	var conf config.Config
 	if err := k.Unmarshal("", &conf); err != nil {
-		logger.Fatal(fmt.Sprintf("error loading config: %v", err))
+		logger.Error(fmt.Sprintf("error loading config: %v", err))
+		return 1
 	}
 
 	logger, err := configureLogger(&conf)
 	if err != nil {
-		logger.Fatal(fmt.Sprintf("error configure logger: %v", err))
+		logger.Error(fmt.Sprintf("error configure logger: %v", err))
+		return 1
 	}
 	defer logger.Sync() //nolint:errcheck
 
@@ -76,51 +82,63 @@ func Execute(version, commit, date string) {
 
 	if err := config.Validate(&conf); err != nil {
 		sl.Error(fmt.Sprintf("error validating config: %v", err))
-		os.Exit(1)
+		return 1
 	}
 
 	oidcClient, err := oauth2.NewProvider(sl, &conf)
 	if err != nil {
 		sl.Error(err.Error())
-		os.Exit(1)
+		return 1
 	}
 
-	oidcClient.OAuthConfig()
-
 	openvpnClient := openvpn.NewClient(sl, &conf)
+	done := make(chan int, 1)
 
 	go func() {
 		stdLogger, err := zap.NewStdLogAt(logger, zap.ErrorLevel)
 		if err != nil {
 			sl.Error(err.Error())
-			os.Exit(1)
+			done <- 1
 		}
 
 		server := &http.Server{
 			Addr:     conf.Http.Listen,
 			ErrorLog: stdLogger,
-			Handler:  oauth2.Handler(sl, &oidcClient, &conf, openvpnClient),
+			Handler:  oauth2.Handler(sl, oidcClient, &conf, openvpnClient),
 		}
 
 		if conf.Http.Tls {
 			sl.Info(fmt.Sprintf("HTTPS server listen on %s with base url %s", conf.Http.Listen, conf.Http.BaseUrl))
 			if err := server.ListenAndServeTLS(conf.Http.CertFile, conf.Http.KeyFile); err != nil {
 				sl.Error(err.Error())
-				os.Exit(1)
+				done <- 1
 			}
 		} else {
 			sl.Info(fmt.Sprintf("HTTP server listen on %s with base url %s", conf.Http.Listen, conf.Http.BaseUrl))
 			if err := server.ListenAndServe(); err != nil {
 				sl.Error(err.Error())
-				os.Exit(1)
+				done <- 1
 			}
 		}
 	}()
 
-	if err := openvpnClient.Connect(); err != nil {
-		sl.Error(err.Error())
-		os.Exit(1)
-	}
+	sigs := make(chan os.Signal, 1)
+	go func() {
+		sig := <-sigs
+		logger.Info(fmt.Sprintf("reciving signal: %s", sig.String()))
+		openvpnClient.Shutdown()
+	}()
+
+	go func() {
+		defer openvpnClient.Shutdown()
+		if err := openvpnClient.Connect(); err != nil {
+			sl.Error(err.Error())
+			done <- 1
+		}
+		done <- 0
+	}()
+
+	return <-done
 }
 
 func configureLogger(conf *config.Config) (*zap.Logger, error) {