+++ internal/utils/slices.go
@@ -1,0 +1,10 @@
+package utils
+
+func CastToSlice[T any](tokenGroups any) []T {
+	tokenGroupsInterfaceList := tokenGroups.([]any)
+	tokenGroupsList := make([]T, len(tokenGroupsInterfaceList))
+	for i, v := range tokenGroupsInterfaceList {
+		tokenGroupsList[i] = v.(T)
+	}
+	return tokenGroupsList
+}