+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2023-09-24
+
+### Added
+
+- deb/rpm packages as binary release
+- Added systemd unit file
+- Added OAUTH2 provider construct for custom provider handling
+
+### Fixes
+
+- openvpn.bypass is nil
+
 ## [1.2.1] - 2023-09-16
 
 ### Changes